Rename app/models/run.py into app/models/job.py (close #391)
