+++ fractal_server/app/models/__init__.py
@@ -1,6 +1,6 @@
 from ...common.schemas import *  # noqa F401
+from .job import *  # noqa: F403, F401
 from .project import *  # noqa: F403, F401
-from .run import *  # noqa: F403, F401
 from .security import *  # noqa: F403, F401
 from .state import State  # noqa: F401
 from .task import *  # noqa: F403, F401
